+++ hpobench/benchmarks/ml/pybnn.py
@@ -12,6 +12,8 @@ Changelog:
 ==========
 0.0.2:
 * Standardize the structure of the meta information
+* The minimum number of burn in steps was allowed to be 0. But then Theano throws an RunTimeError. Limit the number of
+  `burnin_steps` to be at least 1.
 
 0.0.1:
 * First implementation
@@ -107,7 +109,8 @@ class BayesianNeuralNetworkBenchmark(AbstractBenchmark):
         self.rng = rng_helper.get_rng(rng=rng, self_rng=self.rng)
         np.random.seed(self.rng.randint(1, 10000))
 
-        burn_in_steps = int(configuration['burn_in'] * fidelity['budget'])
+        # See comment in objective function test
+        burn_in_steps = max(1, int(configuration['burn_in'] * fidelity['budget']))
 
         net = partial(_get_net,
                       n_units_1=configuration['n_units_1'],
@@ -179,7 +182,9 @@ class BayesianNeuralNetworkBenchmark(AbstractBenchmark):
         self.rng = rng_helper.get_rng(rng=rng, self_rng=self.rng)
         np.random.seed(self.rng.randint(1, 10000))
 
-        burn_in_steps = int(configuration['burn_in'] * fidelity['budget'])
+        # `burn_in_steps` must be at least 1, otherwise, theano will raise an RuntimeError. (Actually, the definition of
+        # the config space allows as lower limit a zero. In this case, set the number of steps to 1.)
+        burn_in_steps = max(1, int(configuration['burn_in'] * fidelity['budget']))
 
         net = partial(_get_net,
                       n_units_1=configuration['n_units_1'],